# appointment-scheduling-headlessAPI-rest/appointment-scheduling-headlessAPI-rest-impl/rest-openapi.yaml
openapi: 3.0.1

info:
  title: "Appointment Scheduling Headless API"
  version: v1.0
  description: >
    Liferay Headless API for managing doctor-patient appointments.
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

components:
  schemas:
    Appointment:
      description: >
        Represents appointment scheduling details between doctors and patients.
      type: object
      properties:
        appointmentId:
          type: integer
          format: int64
        doctorId:
          type: integer
          format: int64
        patientId:
          type: integer
          format: int64
        appointmentDate:
          type: string
          format: date
        timeSlot:
          type: string
        status:
          type: string
          example: "Scheduled"

paths:

  /get-appointments:
    get:
      operationId: getAppointments
      tags: ["Appointment"]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        '403':
          description: Access denied or no appointments found.
        '404':
          description: API not found.

    post:
      operationId: addAppointment
      tags: ["Appointment"]
      summary: "Create a new appointment"
      description: "Add a new doctor-patient appointment."
      requestBody:
        description: "Appointment details to be created."
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        '200':
          description: Appointment added successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        '400':
          description: Invalid input data.
        '403':
          description: Access denied.

  /appointments/{appointmentId}:
    put:
      operationId: updateAppointment
      tags: ["Appointment"]
      summary: "Update an appointment"
      description: "Modify details of an existing appointment."
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: "Updated appointment data."
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        '200':
          description: Appointment updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        '403':
          description: No appointment found or unauthorized.
        '404':
          description: Appointment not found.

    delete:
      operationId: deleteAppointment
      tags: ["Appointment"]
      summary: "Delete an appointment"
      description: "Delete an appointment by its ID."
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Appointment deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Appointment deleted successfully"
        '403':
          description: Access denied or appointment not found.
        '404':
          description: API not found.